model/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0  # Downgrade dari 2.2.2
torchvision==0.16.0

fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
ultralytics==8.0.196  # Pastikan 8.0.196
numpy==1.24.3
opencv-python-headless==4.8.1.78
Pillow==10.1.0
